+++ src/cb_request_executor.c
@@ -9,14 +9,14 @@
 #endif
 
 
-#include "cbor/cb_cbor.h"
-#include "cbor/cb_cbor_int.h"
-#include "cbor/cb_cbor_string.h"
+#include "common/cbor/cb_cbor.h"
+#include "common/cbor/cb_cbor_int.h"
+#include "common/cbor/cb_cbor_string.h"
 #include "cb_msg.h"
 #include "cb_request.h"
 #include "cb_request_executor.h"
 #include "cb_ops.h"
-#include "cb_fs.h"
+#include "common/cb_fs.h"
 
 #include <stdio.h>
 #include <stddef.h>
