+++ src/cli/cb_cli_interpreter.c
@@ -6,7 +6,7 @@
 #include "cli/cb_cli_interpreter.h"
 #include "cb_msg_encode.h"
 #include "cli/cb_cli_help.h"
-#include "cbor/cb_cbor_int.h"
+#include "common/cbor/cb_cbor_int.h"
 
 #include <errno.h>
 #include <stdio.h>
