+++ src/common/cbor/cb_cbor_simple.c
@@ -3,7 +3,7 @@
  *
  */
 
-#include "cbor/cb_cbor_simple.h"
+#include "common/cbor/cb_cbor_simple.h"
 
 size_t cb_cbor_encode_null(uint8_t *ev, size_t size) {
   if (size >= 1) {
