+++ src/common/cbor/cb_cbor_string.c
@@ -5,9 +5,9 @@
 
 #include <string.h>
 
-#include "cbor/cb_cbor_string.h"
-#include "cbor/cb_cbor_int.h"
-#include "cbor/cb_cbor.h"
+#include "common/cbor/cb_cbor_string.h"
+#include "common/cbor/cb_cbor_int.h"
+#include "common/cbor/cb_cbor.h"
 
 //////////////
 // ENCODING //
