+++ tests/test_cb_ops.c
@@ -11,8 +11,8 @@
 #include <fcntl.h>
 #include <string.h>
 
-#include "cb_fs.h"
-#include "cbor/cb_cbor_int.h"
+#include "common/cb_fs.h"
+#include "common/cbor/cb_cbor_int.h"
 #include "cb_ops.h"
 
 int fd_uint;
